moved assert_cant_shield to a new mod
